moving setter/getter funcs to keeper.go, adding tests
